% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{admin_data}
\alias{admin_data}
\alias{bkg_krs}
\alias{bkg_states}
\alias{bkg_germany}
\title{German administrative boundaries}
\format{
For the dataframe format, see \code{\link{bkg_admin}}.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 25 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 25 columns.
}
\source{
© BKG (2025) dl-de/by-2-0, data sources:
\url{https://sgx.geodatenzentrum.de/web_public/gdz/datenquellen/Datenquellen_vg_nuts.pdf}
}
\usage{
bkg_krs

bkg_states

bkg_germany
}
\description{
Three \code{\link[sf:st_sf]{sf}} dataframes containing all geometries of
German districts, federal states, and the country, respectively. The
reference year is 2023.
}
\examples{
bkg_krs
}
\seealso{
\code{\link{bkg_admin}}

Other datasets: 
\code{\link{nuts_data}}
}
\concept{datasets}
\keyword{datasets}
