% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_pois.R
\name{bkg_seaports}
\alias{bkg_seaports}
\title{Seaports}
\usage{
bkg_seaports(
  ...,
  bbox = NULL,
  poly = NULL,
  predicate = "intersects",
  epsg = 3035,
  properties = NULL,
  max = NULL
)
}
\arguments{
\item{...}{Used to construct CQL filters. Dot arguments accept an R-like
syntax that is converted to CQL queries internally. These queries basically
consist of a property name on the left, an aribtrary vector on the right,
and an operator that links both sides. If multiple queries are provided,
they will be chained with \code{AND}. The following operators and their
respective equivalents in CQL are supported:

\tabular{ll}{
\strong{R} \tab \strong{CQL}\cr
\code{==}      \tab \code{=}    \cr
\code{!=}      \tab \code{<>}   \cr
\code{<}       \tab \code{<}    \cr
\code{>}       \tab \code{>}    \cr
\code{>=}      \tab \code{>=}   \cr
\code{<=}      \tab \code{<=}   \cr
\code{~}       \tab \code{~}    \cr
\code{\%!~\%}    \tab \code{!~}   \cr
\code{\%LIKE\%}  \tab \code{LIKE} \cr
\code{\%ILIKE\%} \tab \code{ILIKE}\cr
\code{\%in\%}    \tab \code{IN}
}

To construct more complex queries, you can use the \code{filter} argument
to pass CQL queries directly.}

\item{bbox}{An sf geometry or a boundary box vector of the format
\code{c(xmin, ymin, xmax, ymax)}. Used as a geometric filter to include
only those geometries that relate to \code{bbox} according to the predicate
specified in \code{predicate}.}

\item{poly}{An sf geometry. Used as a geometric filter to include
only those geometries that relate to \code{poly} according to the predicate
specified in \code{predicate}.}

\item{predicate}{A spatial predicate that is used to relate the output
geometries with the object specified in \code{bbox} or \code{poly}. For
example, if \code{predicate = "within"}, and \code{bbox} is specified,
returns only those geometries that lie within \code{bbox}. Can be one of
\code{"equals"}, \code{"disjoint"}, \code{"intersects"}, \code{"touches"},
\code{"crosses"}, \code{"within"}, \code{"contains"}, \code{"overlaps"},
\code{"relate"}, \code{"dwithin"}, or \code{"beyond"}. Defaults to
\code{"intersects"}.}

\item{epsg}{An EPSG code specifying a coordinate reference system of the
output. If you're unsure what this means, try running
\code{sf::st_crs(...)$epsg} on a spatial object that you are working with.
Defaults to 3035.}

\item{properties}{Vector of columns to include in the output.}

\item{max}{Maximum number of results to return.}
}
\value{
A dataframe containing the following columns: \itemize{
\item{\code{name}}: Geographical name of the POI
\item{\code{gemeinde}}: Municipality name
\item{\code{verwaltung}}: Administrative association name
\item{\code{kreis}}: District name
\item{\code{regierungs}}: Government region name
\item{\code{bundesland}}: Federal state name
\item{\code{poi_id}: Unique primary key of a point of interest}
\item{\code{betreiber}: Operator of the seaport}
\item{\code{homepage}: Homepage of the operator}
\item{\code{typ}: Type of seaport. Can be "Seehafen" (seaport) or
"See- und Binnenhafen" (sea and inland port)}
\item{\code{art}: Type of seaport by freight. Can be: \itemize{
\item{Güter: Goods}
\item{Güter und Passagiere: Goods and passengers}
\item{Passagiere: Passengers}
}}
\item{\code{quelle}: Source of the information. Can be: \itemize{
\item{BSH: Federal Maritime and Hydrographic Agency}
\item{MarWiLo: Maritime Wirtschaft & Logistik}
\item{ZDS-Seehäfen: Zentralverband der deutschen Seehafenbetriebe}
}}
}
}
\description{
Retrieve seaports to the North and Baltic Sea in Northern Germany.
}
\examples{
# Get only seaports that co-function as inland ports
ports <- bkg_seaports(typ == "See- und Binnenhafen")
germany <- bkg_admin(level = "sta", scale = "5000", gf == 9)
plot(germany$geometry)
plot(ports$geometry, add = TRUE)
}
\seealso{
\href{https://sgx.geodatenzentrum.de/web_public/gdz/dokumentation/deu/poi-open.pdf}{\code{wfs_poi_open} documentation}

\href{https://mis.bkg.bund.de/trefferanzeige?docuuid=461731F5-62B1-45A9-8E7E-BF6AF93E3EFA}{\code{wfs_poi_open} MIS record}

Other points of interest: 
\code{\link{bkg_airports}()},
\code{\link{bkg_crossings}()},
\code{\link{bkg_heliports}()},
\code{\link{bkg_kilometrage}()},
\code{\link{bkg_stations}()},
\code{\link{bkg_trauma_centers}()}
}
\concept{points of interest}
