% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nuts_data}
\alias{nuts_data}
\alias{bkg_nuts1}
\alias{bkg_nuts2}
\alias{bkg_nuts3}
\title{German NUTS \code{MULTIPOLYGON}s}
\format{
For the dataframe format, see \code{\link{bkg_nuts}}.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 38 rows and 7 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 400 rows and 7 columns.
}
\source{
© BKG (2025) dl-de/by-2-0, data sources:
\url{https://sgx.geodatenzentrum.de/web_public/gdz/datenquellen/datenquellen_vg_nuts.pdf}
}
\usage{
bkg_nuts1

bkg_nuts2

bkg_nuts3
}
\description{
Three \code{\link[sf:st_sf]{sf}} dataframes containing all geometries of
German NUTS1, NUTS2, and NUTS3 regions, respectively. The reference year
is 2023.
}
\examples{
bkg_nuts1
}
\seealso{
\code{\link{bkg_nuts}}

Other datasets: 
\code{\link{admin_data}}
}
\concept{datasets}
\keyword{datasets}
