% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_pois.R
\name{bkg_kilometrage}
\alias{bkg_kilometrage}
\title{Kilometrage}
\usage{
bkg_kilometrage(
  ...,
  bbox = NULL,
  poly = NULL,
  predicate = "intersects",
  epsg = 3035,
  properties = NULL,
  max = NULL
)
}
\value{
A dataframe containing the following columns: \itemize{
\item{\code{name}}: Geographical name of the POI
\item{\code{gemeinde}}: Municipality name
\item{\code{verwaltung}}: Administrative association name
\item{\code{kreis}}: District name
\item{\code{regierungs}}: Government region name
\item{\code{bundesland}}: Federal state name
\item{\code{bez}: Label of the federal motorway}
\item{\code{kilometer}: Kilometrage of the motorway}
\item{\code{richtung}: Direction of the kilometrage}
}
}
\description{
Get kilometrages of German federal motorways. Kilometrages are markers
for each kilometer of a highway.
}
\examples{
\dontshow{if (ffm_run_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get the kilometrage of the A2 motorway
a2 <- bkg_kilometrage(bez == "A2")
plot(a2["kilometer"], pch = 16)
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://sgx.geodatenzentrum.de/web_public/gdz/dokumentation/deu/poi-open.pdf}{\code{wfs_poi_open} documentation}

\href{https://mis.bkg.bund.de/trefferanzeige?docuuid=461731F5-62B1-45A9-8E7E-BF6AF93E3EFA}{\code{wfs_poi_open} MIS record}

The \href{https://cran.r-project.org/package=rLFT}{\code{rLFT}} package for linear referencing

Other points of interest: 
\code{\link{bkg_airports}()},
\code{\link{bkg_crossings}()},
\code{\link{bkg_heliports}()},
\code{\link{bkg_seaports}()},
\code{\link{bkg_stations}()},
\code{\link{bkg_trauma_centers}()}
}
\concept{points of interest}
