% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toponyms.R
\name{bkg_toponyms}
\alias{bkg_toponyms}
\title{Toponyms}
\usage{
bkg_toponyms(..., filter = NULL, epsg = 3035, properties = NULL, max = NULL)
}
\value{
A dataframe containing the following columns:
}
\description{
Get geographic names including toponyms and endonyms.
}
